wearable-maternal-monitor/cloud/requirements.txt
paho-mqtt>=2.1.0
pandas>=2.2.0
scikit-learn>=1.4.0
numpy>=1.26.0
SQLAlchemy>=2.0.29
mysql-connector-python>=9.0.0
python-dotenv>=1.0.1
tqdm>=4.66.0
